refactor(pdf): split pdfHtmlTemplate at the 300-line limit
The fixes in this branch grew it to 504 lines and `pnpm lint:file-size` failed.
Split rather than baselined — a baseline entry records the breakage as
acceptable, which is what the ratchet exists to prevent.

The seam is medium, not size: `pdfPrintCss.ts` holds the rules that exist only
because the output is paper — page-break behaviour, fitting oversized figures to
the sheet, table layout under a fixed width, the forced light palette.
`pdfHtmlTemplate.ts` keeps the document composer. `pdfOptions.ts` carries the
options type and its keyed tables so the CSS builders can take the type without
importing the composer, which would be circular.

pdfHtmlTemplate re-exports the moved names, so every existing importer keeps one
entry point and no call site changed.

198 / 294 / 43 lines. Verified: typecheck clean, 829 tests across 48 files pass,
file-size gate green.

Refs #1284
